@@ -5,12 +5,13 @@
 #------------------------------------------------------------------------------------------#
 """
 """
+import warnings
 from os.path import join, dirname
 
 from textx.metamodel import metamodel_from_file
 
 from pyccel.parser.syntax.basic import BasicStmt
-from pyccel.ast.headers   import FunctionHeader, MethodHeader, VariableHeader, Template
+from pyccel.ast.headers   import FunctionHeader, MethodHeader, Template
 from pyccel.ast.headers   import MetaVariable , UnionType, InterfaceHeader
 from pyccel.ast.headers   import construct_macro, MacroFunction, MacroVariable
 from pyccel.ast.core      import FunctionDefArgument, EmptyNode
@@ -19,7 +20,8 @@ from pyccel.ast.datatypes import dtype_and_precision_registry as dtype_registry,
 from pyccel.ast.datatypes import NativeNumeric
 from pyccel.ast.literals  import LiteralString, LiteralInteger, LiteralFloat
 from pyccel.ast.literals  import LiteralTrue, LiteralFalse
-from pyccel.ast.internals import PyccelSymbol
+from pyccel.ast.internals import PyccelSymbol, AnnotatedPyccelSymbol
+from pyccel.ast.type_annotations import SyntacticTypeAnnotation
 from pyccel.errors.errors import Errors
 from pyccel.utilities.stage import PyccelStage
 
@@ -62,58 +64,43 @@ class FuncType(BasicStmt):
         return d_var
 
 class TemplateStmt(BasicStmt):
-    """Base class representing a  template in the grammar."""
-    def __init__(self, **kwargs):
-        self.dtypes  = kwargs.pop('dtypes')
-        self.name   = kwargs.pop('name')
-        BasicStmt.__init__(self)
+    """
+    Base class representing a template in the grammar.
+
+    Base class representing a template in the grammar.
+    To be removed when header support is deprecated.
+
+    Parameters
+    ----------
+    name : str
+        The name of the template type symbol.
+    dtypes : list of str
+        A list of the types that the template describes.
+    **kwargs : dict
+        The textx arguments.
+    """
+    def __init__(self, *, name, dtypes, **kwargs):
+        self.dtypes = dtypes
+        self.name   = name
+        super().__init__(**kwargs)
 
     @property
     def expr(self):
+        """
+        Get the Pyccel equivalent of this object.
+
+        Get the Pyccel equivalent of this object.
+        """
         if any(isinstance(d_type, FuncType) for d_type in self.dtypes):
             msg = 'Functions in a template are not supported yet'
             errors.report(msg,
                         severity='error')
             return EmptyNode()
 
-        possible_dtypes = {tuple(t.expr.items())  for t in self.dtypes}
-        dtypes = tuple(dict(d_type) for d_type in possible_dtypes)
+        dtypes = {SyntacticTypeAnnotation.build_from_textx(t)  for t in self.dtypes}
         return Template(self.name, dtypes)
 
-class ListType(BasicStmt):
-    """Base class representing a  ListType in the grammar."""
-
-    def __init__(self, **kwargs):
-        """
-        Constructor for a TypeHeader.
-
-        dtype: list of str
-        """
-        self.dtype = kwargs.pop('dtype')
-
-        super(ListType, self).__init__(**kwargs)
-
-    @property
-    def expr(self):
-        dtypes = [str(i.expr['datatype']) for i in self.dtype]
-        precisions = [i.expr['precision'] for i in self.dtype]
-        if not (all(dtypes[0]==i for i in dtypes)):
-            raise TypeError('all element of the TypeList must have the same type')
-
-        d_var = {}
-        d_var['datatype'] = str(dtypes[0])
-        d_var['rank'] = len(dtypes)
-        d_var['memory_handling'] = 'alias' if len(dtypes) > 0 else 'stack'
-        d_var['precision'] = max(precisions)
-        d_var['order'] = 'C'
-        d_var['is_func'] = False
-        d_var['is_const'] = False
-        if not(d_var['precision']):
-            if d_var['datatype'] in ['double','float','complex','int']:
-                d_var['precision'] = default_precision[d_var['datatype']]
-        return d_var
-
-class Type(BasicStmt):
+class VariableType(BasicStmt):
     """
     Base class representing a header type in the grammar.
 
@@ -124,7 +111,7 @@ class Type(BasicStmt):
     dtype : str
         The variable type.
 
-    prec : int
+    prec : int, optional
         The precision of the object.
 
     trailer : iterable, TrailerSubscriptsList
@@ -136,47 +123,15 @@ class Type(BasicStmt):
         The textx arguments.
     """
 
-    def __init__(self, dtype, prec, trailer = (), **kwargs):
+    def __init__(self, dtype, prec = None, trailer = (), **kwargs):
         self.dtype   = dtype
         self.prec    = prec
-        self.trailer = trailer
-
-        super(Type, self).__init__(**kwargs)
-
-    @property
-    def expr(self):
-        """
-        Get the dictionary describing the type.
-
-        Get the dictionary describing the type.
-        """
-        dtype = self.dtype
-        precision = self.prec
-        if dtype in dtype_registry.keys():
-            dtype,precision = dtype_registry[dtype]
-        trailer = self.trailer
-        order = 'C'
-
         if trailer:
-            if trailer.order:
-                order = str(trailer.order)
-            trailer = [str(i) for i in trailer.args]
+            self.trailer = trailer
         else:
-            trailer = []
-        d_var={}
-        d_var['datatype']=dtype
-        d_var['rank'] = len(trailer)
-        d_var['memory_handling'] = 'heap' if len(trailer) > 0 else 'stack'
-        d_var['precision']  = precision
-        d_var['is_func'] = False
-        d_var['is_const'] = False
-        if not(precision):
-            if dtype in NativeNumeric:
-                d_var['precision'] = default_precision[dtype]
+            self.trailer = []
 
-        if d_var['rank']>1:
-            d_var['order'] = order
-        return d_var
+        super().__init__(**kwargs)
 
 class ShapedID(BasicStmt):
     """class representing a ShapedID in the grammar.
@@ -205,30 +160,58 @@ class ShapedID(BasicStmt):
 
         return d_var
 
-class TypeHeader(BasicStmt):
-    pass
-
 class StringStmt(BasicStmt):
-    def __init__(self, **kwargs):
-        self.arg = kwargs.pop('arg')
+    """
+    Class describing a string in a macro.
+
+    Class describing a string in a macro.
+    To be removed when macro support is deprecated.
+
+    Parameters
+    ----------
+    arg : str
+        The string.
+    **kwargs : dict
+        TextX keyword arguments.
+    """
+    def __init__(self, arg, **kwargs):
+        self.arg = arg
+        super().__init__(**kwargs)
     @property
     def expr(self):
         return LiteralString(str(self.arg))
 
 class UnionTypeStmt(BasicStmt):
-    def __init__(self, **kwargs):
-        """
-        Constructor for a TypeHeader.
+    """
+    Class describing a union of possible types.
 
-        dtype: list fo str
-        """
-        self.dtypes = kwargs.pop('dtypes')
-        self.const = kwargs.pop('const')
+    A class object describing a union of possible types described in a type descriptor.
+    These types are either VariableTypes or FuncTypes.
 
-        super(UnionTypeStmt, self).__init__(**kwargs)
+    Parameters
+    ----------
+    dtypes : list of VariableHeader | FuncHeader
+        A list of the possible types described.
+    const : bool
+        A boolean indicating if the generated object will be constant or
+        modifiable.
+    **kwargs : dict
+        TextX keyword arguments.
+    """
+    def __init__(self, dtypes, const = None, **kwargs):
+        self.dtypes = list(dtypes)
+        self.const = const
+
+        super().__init__(**kwargs)
 
     @property
     def expr(self):
+        """
+        Get the Pyccel equivalent of this object.
+
+        Get the Pyccel equivalent of this object.
+        To be removed when header support is deprecated.
+        """
         dtypes = [i.expr for i in self.dtypes]
         if self.const:
             for d_type in dtypes:
@@ -246,101 +229,131 @@ class UnionTypeStmt(BasicStmt):
         return UnionType(dtypes)
 
 class HeaderResults(BasicStmt):
-    """Base class representing a HeaderResults in the grammar."""
+    """
+    Base class representing a HeaderResults in the grammar.
 
-    def __init__(self, **kwargs):
-        """
-        Constructor for a HeaderResults.
+    Base class representing a HeaderResults in the grammar.
+    To be removed when header support is deprecated.
 
-        decs: list of TypeHeader
-        """
-        self.decs = kwargs.pop('decs')
+    Parameters
+    ----------
+    decs : list of VariableHeader | FuncHeader
+        List of TypeHeaders defining the return type(s).
+    **kwargs : dict
+        TextX keyword arguments.
+    """
 
-        super(HeaderResults, self).__init__(**kwargs)
+    def __init__(self, decs, **kwargs):
+        self.decs = decs
+
+        super().__init__(**kwargs)
 
     @property
     def expr(self):
-        decs = [i.expr for i in self.decs]
+        """
+        Get the Pyccel equivalent of this object.
+
+        Get the Pyccel equivalent of this object.
+        """
+        decs = SyntacticTypeAnnotation.build_from_textx(self.decs)
         return decs
 
 
 class VariableHeaderStmt(BasicStmt):
-    """Base class representing a header statement in the grammar."""
+    """
+    Base class representing a header statement in the grammar.
 
-    def __init__(self, **kwargs):
-        """
-        Constructor for a VariableHeader statement.
-        In the case of builtin datatypes, we export a Variable
+    Base class representing a header statement in the grammar.
+    To be removed when header support is deprecated.
 
-        name: str
-            variable name
-        dec: list, tuple
-            list of argument types
-        """
-        self.name = kwargs.pop('name')
-        self.dec  = kwargs.pop('dec')
+    Parameters
+    ----------
+    name : str
+        Variable name.
+    dec : list, tuple
+        List of argument types.
+    **kwargs : dict
+        TextX keyword arguments.
+    """
+
+    def __init__(self, name, dec, **kwargs):
+        self.name = name
+        self.dec  = dec
 
         super(VariableHeaderStmt, self).__init__(**kwargs)
 
     @property
     def expr(self):
-        dtype = self.dec.expr
+        """
+        Convert a VariableHeaderStmt generated by TextX to a Pyccel AST object.
 
-        return VariableHeader(self.name, dtype)
+        Convert a VariableHeaderStmt generated by TextX to a Pyccel AST object.
+        """
+        warnings.warn("Support for specifying types via headers will be removed in " +
+                      "a future version of Pyccel. This annotation may be unnecessary " +
+                      "in your code. If you find it is necessary please open a discussion " +
+                      "at https://github.com/pyccel/pyccel/discussions so we do not " +
+                      "remove support until an alternative is in place.", FutureWarning)
+        dtype = SyntacticTypeAnnotation.build_from_textx(self.dec)
+
+        return AnnotatedPyccelSymbol(self.name, annotation=dtype)
 
 class FunctionHeaderStmt(BasicStmt):
-    """Base class representing a function header statement in the grammar."""
+    """
+    Base class representing a function header statement in the grammar.
 
-    def __init__(self, **kwargs):
-        """
-        Constructor for a FunctionHeader statement
+    Base class representing a function header statement in the grammar.
+    To be removed when header support is deprecated.
 
-        name: str
-            function name
-        kind: str
-            one among {function, method}
-        decs: list, tuple
-            list of argument types
-        results: list, tuple
-            list of output types
-        """
-        self.name = kwargs.pop('name')
-        self.kind = kwargs.pop('kind', None)
-        self.static = kwargs.pop('static', None)
-        self.decs = kwargs.pop('decs')
-        self.results = kwargs.pop('results', None)
+    Parameters
+    ----------
+    name : str
+        Function name.
+    decs : list, tuple
+        List of argument types.
+    kind : str
+        One among {function, method}.
+    static : bool
+        Indicates if the function is a static function.
+    results : list, tuple
+        List of output types.
+    **kwargs : dict
+        Arguments defined by textx.
+    """
 
-        super(FunctionHeaderStmt, self).__init__(**kwargs)
+    def __init__(self, name, decs, kind=None, static = None, results=None, **kwargs):
+        self.name = name
+        self.kind = kind
+        self.static = static
+        self.decs = decs
+        self.results = results
+
+        super().__init__(**kwargs)
 
     @property
     def expr(self):
-        # TODO: do we need dtypes and results to be attributs of the class?
-        dtypes = []
-        for dec in self.decs:
-            if isinstance(dec,UnionTypeStmt):
-                dtypes += [dec.expr]
+        """
+        Get the Pyccel equivalent of this object.
+
+        Get the Pyccel equivalent of this object.
+        """
+        dtypes = SyntacticTypeAnnotation.build_from_textx(self.decs)
 
         if self.kind is None:
             kind = 'function'
         else:
             kind = str(self.kind)
 
-        is_static = False
-        if self.static == 'static':
-            is_static = True
+        is_static = self.static
 
         results = []
         if self.results:
             results = self.results.expr
 
         if kind == 'method':
-            dtype = dtypes[0]
-            if isinstance(dtype, UnionType):
-                cls_instance = dtype.args[0]['datatype']
-            else:
-                cls_instance = dtype['datatype']
-            dtypes = dtypes[1:] # remove the attribut
-            return MethodHeader((cls_instance, self.name), dtypes, [] )
+            return MethodHeader(self.name, dtypes,
+                                  results=results,
+                                  is_static=is_static)
         else:
             return FunctionHeader(self.name,
                                   dtypes,
@@ -539,8 +552,7 @@ class FunctionMacroStmt(BasicStmt):
 #################################################
 # whenever a new rule is added in the grammar, we must update the following
 # lists.
-hdr_classes = [Header, TypeHeader,
-               Type, ListType, UnionTypeStmt, FuncType,
+hdr_classes = [Header, VariableType, UnionTypeStmt, FuncType,
                ShapedID,
                HeaderResults,
                FunctionHeaderStmt,
@@ -557,8 +569,10 @@ this_folder = dirname(__file__)
 
 # Get meta-model from language description
 grammar = join(this_folder, '../grammar/headers.tx')
+types_grammar = join(this_folder, '../grammar/types.tx')
 
 meta = metamodel_from_file(grammar, classes=hdr_classes)
+types_meta = metamodel_from_file(types_grammar)
 
 def parse(filename=None, stmts=None):
     """ Parse header pragmas
