@@ -14,12 +14,12 @@ from pyccel.errors.errors import PyccelError
 
 from pyccel.utilities.stage import PyccelStage
 
-from .basic     import Basic, PyccelAstNode
+from .basic     import PyccelAstNode, TypedAstNode
 from .datatypes import (NativeInteger, NativeBool, NativeFloat,
                         NativeComplex, NativeString, NativeGeneric)
 from .internals import PyccelInternalFunction, max_precision, Slice
 from .literals  import LiteralInteger, LiteralFloat, LiteralComplex, Nil
-from .literals  import Literal, LiteralImaginaryUnit, get_default_literal_value
+from .literals  import Literal, LiteralImaginaryUnit, convert_to_literal
 from .literals  import LiteralString
 from .operators import PyccelAdd, PyccelAnd, PyccelMul, PyccelIsNot
 from .operators import PyccelMinus, PyccelUnarySub, PyccelNot
@@ -119,6 +119,7 @@ class PythonImag(PythonComplexProperty):
     """
     Represents a call to the .imag property.
 
+    Represents a call to the .imag property of an object with a complex type.
     e.g:
     >>> a = 1+2j
     >>> a.imag
@@ -126,14 +127,14 @@ class PythonImag(PythonComplexProperty):
 
     Parameters
     ----------
-    arg : PyccelAstNode
-        The object which the property is called from.
+    arg : TypedAstNode
+        The object on which the property is called.
     """
     __slots__ = ()
     name = 'imag'
     def __new__(cls, arg):
         if arg.dtype is not NativeComplex():
-            return get_default_literal_value(arg.dtype)
+            return convert_to_literal(0, dtype = arg.dtype)
         else:
             return super().__new__(cls)
 
@@ -155,7 +156,7 @@ class PythonConjugate(PyccelInternalFunction):
 
     Parameters
     ----------
-    arg : PyccelAstNode
+    arg : TypedAstNode
         The variable/expression which was passed to the
         conjugate function.
     """
@@ -187,7 +188,7 @@ class PythonConjugate(PyccelInternalFunction):
         return f'Conjugate({self.internal_var})'
 
 #==============================================================================
-class PythonBool(PyccelAstNode):
+class PythonBool(TypedAstNode):
     """
     Represents a call to Python's native `bool()` function.
 
@@ -228,7 +229,7 @@ class PythonBool(PyccelAstNode):
         return f'Bool({self.arg})'
 
 #==============================================================================
-class PythonComplex(PyccelAstNode):
+class PythonComplex(TypedAstNode):
     """
     Represents a call to Python's native `complex()` function.
 
@@ -342,7 +343,7 @@ class PythonComplex(PyccelAstNode):
         return f"complex({self.real}, {self.imag})"
 
 #==============================================================================
-class PythonEnumerate(Basic):
+class PythonEnumerate(PyccelAstNode):
     """
     Represents a call to Python's native `enumerate()` function.
 
@@ -362,7 +363,7 @@ class PythonEnumerate(Basic):
 
     def __init__(self, arg, start = None):
         if pyccel_stage != "syntactic" and \
-                not isinstance(arg, PyccelAstNode):
+                not isinstance(arg, TypedAstNode):
             raise TypeError('Expecting an arg of valid type')
         self._element = arg
         self._start   = start or LiteralInteger(0)
@@ -389,7 +390,7 @@ class PythonEnumerate(Basic):
         return PythonLen(self.element)
 
 #==============================================================================
-class PythonFloat(PyccelAstNode):
+class PythonFloat(TypedAstNode):
     """
     Represents a call to Python's native `float()` function.
 
@@ -429,7 +430,7 @@ class PythonFloat(PyccelAstNode):
         return f'float({self.arg})'
 
 #==============================================================================
-class PythonInt(PyccelAstNode):
+class PythonInt(TypedAstNode):
     """
     Represents a call to Python's native `int()` function.
 
@@ -466,16 +467,18 @@ class PythonInt(PyccelAstNode):
         return self._arg
 
 #==============================================================================
-class PythonTuple(PyccelAstNode):
+class PythonTuple(TypedAstNode):
     """
-    Represents a call to Python's native tuple() function.
+    Class representing a call to Python's native tuple() function.
 
-    Represents a call to Python's native tuple() function.
+    Class representing a call to Python's native tuple() function
+    which either converts an object to a tuple or initialises a
+    literal tuple.
 
     Parameters
     ----------
-    *args : tuple of PyccelAstNodes
-        The elements of the tuple.
+    *args : tuple of TypedAstNode
+        The arguments passed to the tuple function.
     """
     __slots__ = ('_args','_inconsistent_shape','_is_homogeneous',
             '_dtype','_precision','_rank','_shape','_order')
@@ -537,11 +540,14 @@ class PythonTuple(PyccelAstNode):
                 self._rank  = len(self._shape)
 
         else:
-            self._rank      = max(a.rank for a in args) + 1
+            max_rank = max(a.rank for a in args)
+            self._rank      = max_rank + 1
             self._dtype     = NativeGeneric()
             self._precision = 0
             if self._rank == 1:
                 self._shape     = (LiteralInteger(len(args)), )
+            elif any(a.rank != max_rank for a in args):
+                self._shape     = (LiteralInteger(len(args)), ) + (None,)*(self._rank-1)
             else:
                 self._shape     = (LiteralInteger(len(args)), ) + args[0].shape
 
@@ -647,7 +653,7 @@ class PythonList(PythonTuple):
     __slots__ = ()
 
 #==============================================================================
-class PythonMap(Basic):
+class PythonMap(PyccelAstNode):
     """ Represents the map stmt
     """
     __slots__ = ('_func','_func_args')
@@ -681,11 +687,11 @@ class PythonMap(Basic):
         return PythonLen(self.func_args)
 
 #==============================================================================
-class PythonPrint(Basic):
+class PythonPrint(PyccelAstNode):
 
     """Represents a print function in the code.
 
-    expr : PyccelAstNode
+    expr : TypedAstNode
         The expression to print
     file: String (Optional)
         Select 'stdout' (default) or 'stderr' to print to
@@ -719,7 +725,7 @@ class PythonPrint(Basic):
         return self._file
 
 #==============================================================================
-class PythonRange(Basic):
+class PythonRange(PyccelAstNode):
 
     """
     Represents a range.
@@ -849,7 +855,7 @@ class PythonSum(PyccelInternalFunction):
     _order = None
 
     def __init__(self, arg):
-        if not isinstance(arg, PyccelAstNode):
+        if not isinstance(arg, TypedAstNode):
             raise TypeError(f'Unknown type of {type(arg)}.' )
         self._dtype = arg.dtype
         self._precision = -1
@@ -928,7 +934,7 @@ class PythonMin(PyccelInternalFunction):
         super().__init__(x)
 
 #==============================================================================
-class Lambda(Basic):
+class Lambda(PyccelAstNode):
     """
     Represents a call to Python's lambda for temporary functions.
 
@@ -938,7 +944,7 @@ class Lambda(Basic):
     ----------
     variables : tuple of symbols
         The arguments to the lambda expression.
-    expr : PyccelAstNode
+    expr : TypedAstNode
         The expression carried out when the lambda function is called.
     """
     __slots__ = ('_variables', '_expr')
@@ -973,7 +979,7 @@ class Lambda(Basic):
         return f"{self.variables} -> {self.expr}"
 
 #==============================================================================
-class PythonType(Basic):
+class PythonType(PyccelAstNode):
     """
     Represents a call to the Python builtin `type` function.
 
@@ -986,14 +992,14 @@ class PythonType(Basic):
 
     Parameters
     ==========
-    obj : PyccelAstNode
+    obj : TypedAstNode
           The object whose type we wish to investigate.
     """
     __slots__ = ('_dtype','_precision','_obj')
     _attribute_nodes = ('_obj',)
 
     def __init__(self, obj):
-        if not isinstance (obj, PyccelAstNode):
+        if not isinstance (obj, TypedAstNode):
             raise PyccelError(f"Python's type function is not implemented for {type(obj)} object")
         self._dtype = obj.dtype
         self._precision = obj.precision
@@ -1044,7 +1050,8 @@ python_builtin_datatypes_dict = {
     'bool'   : PythonBool,
     'float'  : PythonFloat,
     'int'    : PythonInt,
-    'complex': PythonComplex
+    'complex': PythonComplex,
+    'str'    : LiteralString
 }
 
 def python_builtin_datatype(name):
