@@ -11,7 +11,7 @@ between Python code and C code (using Python/C Api and cwrapper.c).
 from ..errors.errors import Errors
 from ..errors.messages import PYCCEL_RESTRICTION_TODO
 
-from .basic     import Basic, PyccelAstNode
+from .basic     import PyccelAstNode, TypedAstNode
 
 from .datatypes import DataType, default_precision
 from .datatypes import NativeInteger, NativeFloat, NativeComplex
@@ -72,7 +72,7 @@ class PyccelPyObject(DataType):
 #-------------------------------------------------------------------
 
 #TODO: Is there an equivalent to static so this can be a static list of strings?
-class PyArgKeywords(Basic):
+class PyArgKeywords(PyccelAstNode):
     """
     Represents the list containing the names of all arguments to a function.
     This information allows the function to be called by keyword
@@ -106,7 +106,7 @@ class PyArgKeywords(Basic):
         return self._arg_names
 
 #-------------------------------------------------------------------
-class PyArg_ParseTupleNode(Basic):
+class PyArg_ParseTupleNode(PyccelAstNode):
     """
     Represents a call to the function `PyArg_ParseTupleNode`.
 
@@ -198,7 +198,7 @@ class PyArg_ParseTupleNode(Basic):
         return self._arg_names
 
 #-------------------------------------------------------------------
-class PyBuildValueNode(PyccelAstNode):
+class PyBuildValueNode(TypedAstNode):
     """
     Represents a call to the function PyBuildValueNode.
 
@@ -236,7 +236,7 @@ class PyBuildValueNode(PyccelAstNode):
         return self._result_args
 
 #-------------------------------------------------------------------
-class PyModule_AddObject(PyccelAstNode):
+class PyModule_AddObject(TypedAstNode):
     """
     Represents a call to the PyModule_AddObject function.
 
