@@ -13,7 +13,7 @@ from pyccel.utilities.stage import PyccelStage
 
 from ..errors.errors        import Errors, PyccelSemanticError
 
-from .basic                 import PyccelAstNode
+from .basic                 import TypedAstNode
 
 from .datatypes             import (NativeBool, NativeInteger, NativeFloat,
                                     NativeComplex, NativeString,
@@ -122,7 +122,7 @@ def broadcast(shape_1, shape_2):
 
 #==============================================================================
 
-class PyccelOperator(PyccelAstNode):
+class PyccelOperator(TypedAstNode):
     """
     Abstract superclass for all builtin operators.
     The __init__ function is common
@@ -230,7 +230,7 @@ class PyccelUnaryOperator(PyccelOperator):
 
     Parameters
     ----------
-    arg : PyccelAstNode
+    arg : TypedAstNode
         The argument passed to the operator.
     """
     __slots__ = ('_dtype', '_precision','_shape','_rank','_order')
@@ -270,7 +270,7 @@ class PyccelUnary(PyccelUnaryOperator):
 
     Parameters
     ----------
-    arg : PyccelAstNode
+    arg : TypedAstNode
         The argument passed to the operator.
     """
     __slots__ = ()
@@ -295,7 +295,7 @@ class PyccelUnarySub(PyccelUnary):
 
     Parameters
     ----------
-    arg : PyccelAstNode
+    arg: TypedAstNode
         The argument passed to the operator.
     """
     __slots__ = ()
@@ -315,7 +315,7 @@ class PyccelNot(PyccelUnaryOperator):
 
     Parameters
     ----------
-    arg : PyccelAstNode
+    arg : TypedAstNode
         The argument passed to the operator.
     """
     __slots__ = ()
@@ -352,7 +352,7 @@ class PyccelAssociativeParenthesis(PyccelUnaryOperator):
 
     Parameters
     ----------
-    arg : PyccelAstNode
+    arg : TypedAstNode
         The argument in the PyccelAssociativeParenthesis.
     """
     __slots__ = () # ok
@@ -371,9 +371,9 @@ class PyccelBinaryOperator(PyccelOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ('_dtype','_precision','_shape','_rank','_order')
@@ -478,9 +478,9 @@ class PyccelArithmeticOperator(PyccelBinaryOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -503,9 +503,9 @@ class PyccelPow(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -544,9 +544,9 @@ class PyccelAdd(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     simplify : bool
         True if the expression should be simplified to be as compact/readable as
@@ -607,9 +607,9 @@ class PyccelMul(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     simplify : bool
         True if the expression should be simplified to be as compact/readable as
@@ -657,9 +657,9 @@ class PyccelMinus(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     simplify : bool
         True if the expression should be simplified to be as compact/readable as
@@ -707,9 +707,9 @@ class PyccelDiv(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     simplify : bool
         True if the expression should be simplified to be as compact/readable as
@@ -750,9 +750,9 @@ class PyccelMod(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -775,9 +775,9 @@ class PyccelFloorDiv(PyccelArithmeticOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -794,9 +794,9 @@ class PyccelComparisonOperator(PyccelBinaryOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -821,9 +821,9 @@ class PyccelEq(PyccelComparisonOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     simplify : bool
         True if the expression should be simplified to be as compact/readable as
@@ -855,9 +855,9 @@ class PyccelNe(PyccelComparisonOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     simplify : bool
         True if the expression should be simplified to be as compact/readable as
@@ -889,9 +889,9 @@ class PyccelLt(PyccelComparisonOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -911,9 +911,9 @@ class PyccelLe(PyccelComparisonOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -933,9 +933,9 @@ class PyccelGt(PyccelComparisonOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -955,9 +955,9 @@ class PyccelGe(PyccelComparisonOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -973,9 +973,9 @@ class PyccelBooleanOperator(PyccelOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     dtype = NativeBool()
@@ -1009,9 +1009,9 @@ class PyccelAnd(PyccelBooleanOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
@@ -1038,9 +1038,9 @@ class PyccelOr(PyccelBooleanOperator):
 
     Parameters
     ----------
-    arg1 : PyccelAstNode
+    arg1 : TypedAstNode
         The first argument passed to the operator.
-    arg2 : PyccelAstNode
+    arg2 : TypedAstNode
         The second argument passed to the operator.
     """
     __slots__ = ()
