@@ -7,7 +7,7 @@ import functools
 from itertools import chain
 import re
 
-from pyccel.ast.basic     import ScopedNode
+from pyccel.ast.basic     import ScopedAstNode
 
 from pyccel.ast.builtins  import PythonRange, PythonComplex
 from pyccel.ast.builtins  import PythonPrint, PythonType
@@ -343,7 +343,7 @@ class CCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        a : PyccelAstNode
+        a : TypedAstNode
             The object whose storage we are enquiring about.
 
         Returns
@@ -375,7 +375,7 @@ class CCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        expr : PyccelAstNode
+        expr : TypedAstNode
             The Assign Node used to get the lhs and rhs.
 
         Returns
@@ -468,7 +468,7 @@ class CCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        expr : PyccelAstNode
+        expr : TypedAstNode
             The Assign Node used to get the lhs and rhs.
 
         Returns
@@ -499,7 +499,7 @@ class CCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        expr : PyccelAstNode
+        expr : TypedAstNode
             The Assign Node used to get the lhs and rhs.
 
         Returns
@@ -555,7 +555,7 @@ class CCodePrinter(CodePrinter):
         body = func.body
 
         for b in body.body:
-            if isinstance(b, ScopedNode):
+            if isinstance(b, ScopedAstNode):
                 b.scope.update_parent_scope(self.scope, is_loop=True)
 
         # Print any arguments using the same inline function
@@ -621,7 +621,7 @@ class CCodePrinter(CodePrinter):
                 self.scope.insert_symbol(v.name)
 
         for b in body.body:
-            if isinstance(b, ScopedNode):
+            if isinstance(b, ScopedAstNode):
                 b.scope.update_parent_scope(func.scope, is_loop=True)
 
         return code
@@ -962,7 +962,7 @@ class CCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        var : PyccelAstNode
+        var : TypedAstNode
             The object which will be printed.
 
         Returns
@@ -1367,7 +1367,7 @@ class CCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        expr : PyccelAstNode
+        expr : TypedAstNode
             The expression to be cast.
         dtype : Datatype
             The target type of the cast.
@@ -1500,6 +1500,9 @@ class CCodePrinter(CodePrinter):
         variable_address = self._print(ObjectAddress(expr.variable))
         if expr.variable.is_alias:
             return f'free_pointer({variable_address});\n'
+        if isinstance(expr.variable.dtype, CustomDataType):
+            Pyccel__del = expr.variable.cls_base.scope.find('__del__').name
+            return f"{Pyccel__del}({variable_address});\n"
         return f'free_array({variable_address});\n'
 
     def _print_Slice(self, expr):
@@ -2286,6 +2289,9 @@ class CCodePrinter(CodePrinter):
 
     def _print_CustomDataType(self, expr):
         return "struct " + expr.name
+
+    def _print_Del(self, expr):
+        return ''.join(self._print(var) for var in expr.variables)
 
     def _print_ClassDef(self, expr):
         methods = ''.join(self._print(method) for method in expr.methods)
