@@ -123,7 +123,7 @@ class PythonCodePrinter(CodePrinter):
 
         Returns
         -------
-        body : PyccelAstNode
+        body : TypedAstNode
             The expression inside the for loops.
         iterables : list of Iterables
             The iterables over which the for loops iterate.
@@ -197,7 +197,7 @@ class PythonCodePrinter(CodePrinter):
 
         Parameters
         ----------
-        expr : PyccelAstNode
+        expr : TypedAstNode
             The expression whose datatype is being determined.
 
         init_dtype : PythonType, PyccelFunctionDef, LiteralString, str
@@ -1097,7 +1097,7 @@ class PythonCodePrinter(CodePrinter):
         return f"{cls_variable} = {cls_name}({args})\n"
 
     def _print_Del(self, expr):
-        return ''.join(f'del {var}\n' for var in expr.variables)
+        return ''.join(f'del {var.variable}\n' for var in expr.variables)
 
     #------------------OmpAnnotatedComment Printer------------------
 
@@ -1116,6 +1116,25 @@ class PythonCodePrinter(CodePrinter):
         omp_expr = str(expr.txt)
         omp_expr = f'#$omp {omp_expr}\n'
         return omp_expr
+
+    #------------------Annotation Printer------------------
+
+    def _print_UnionTypeAnnotation(self, expr):
+        types = [self._print(t)[1:-1] for t in expr.type_list]
+        return "'" + ' | '.join(types) + "'"
+
+    def _print_SyntacticTypeAnnotation(self, expr):
+        dtypes = expr.dtypes
+        ranks = [f"[{','.join(':'*r)}]" if r>0 else '' for r in expr.ranks]
+        order = [f"(order={o})" if o else '' for o in expr.orders]
+        annot = [f'{d}{r}{o}' for d,r,o in zip(dtypes, ranks, order)]
+        const = 'const ' if expr.is_const else ''
+        return "'" + const + ' | '.join(annot) + "'"
+
+    def _print_FunctionTypeAnnotation(self, expr):
+        args = ', '.join(self._print(a.annotation)[1:-1] for a in expr.args)
+        results = ', '.join(self._print(r.annotation)[1:-1] for r in expr.results)
+        return "'" + f"({results})({args})" + "'"
 
 #==============================================================================
 def pycode(expr, assign_to=None, **settings):
