@@ -67,12 +67,12 @@ class Wrapper:
 
         Parameters
         ----------
-        expr : pyccel.ast.basic.Basic
+        expr : pyccel.ast.basic.PyccelAstNode
             The expression that should be wrapped.
 
         Returns
         -------
-        pyccel.ast.basic.Basic
+        pyccel.ast.basic.PyccelAstNode
             The AST which describes the object that lets you
             access the expression.
         """
@@ -87,12 +87,12 @@ class Wrapper:
 
         Parameters
         ----------
-        expr : pyccel.ast.basic.Basic
+        expr : pyccel.ast.basic.PyccelAstNode
             The expression that should be wrapped.
 
         Returns
         -------
-        pyccel.ast.basic.Basic
+        pyccel.ast.basic.PyccelAstNode
             The AST which describes the object that lets you
             access the expression.
         """
