@@ -10,13 +10,11 @@ import re
 import ast
 import warnings
 
-#==============================================================================
-
-from sympy.core import cache
+from textx.exceptions import TextXSyntaxError
 
 #==============================================================================
 
-from pyccel.ast.basic import Basic
+from pyccel.ast.basic import PyccelAstNode
 
 from pyccel.ast.core import FunctionCall, FunctionCallArgument
 from pyccel.ast.core import Module
@@ -55,21 +53,23 @@ from pyccel.ast.numpyext  import NumpyMatmul
 
 from pyccel.ast.builtins import PythonTuple, PythonList
 from pyccel.ast.builtins import PythonPrint, Lambda
-from pyccel.ast.headers  import MetaVariable
+from pyccel.ast.headers  import MetaVariable, FunctionHeader, MethodHeader
 from pyccel.ast.literals import LiteralInteger, LiteralFloat, LiteralComplex
 from pyccel.ast.literals import LiteralFalse, LiteralTrue, LiteralString
 from pyccel.ast.literals import Nil
 from pyccel.ast.functionalexpr import FunctionalSum, FunctionalMax, FunctionalMin, GeneratorComprehension, FunctionalFor
 from pyccel.ast.variable  import DottedName
 
-from pyccel.ast.internals import Slice, PyccelSymbol, PyccelInternalFunction
+from pyccel.ast.internals import Slice, PyccelSymbol, PyccelInternalFunction, AnnotatedPyccelSymbol
+
+from pyccel.ast.type_annotations import SyntacticTypeAnnotation, UnionTypeAnnotation
 
 from pyccel.parser.base        import BasicParser
 from pyccel.parser.extend_tree import extend_tree
 from pyccel.parser.utilities   import read_file
 from pyccel.parser.utilities   import get_default_path
 
-from pyccel.parser.syntax.headers import parse as hdr_parse
+from pyccel.parser.syntax.headers import parse as hdr_parse, types_meta
 from pyccel.parser.syntax.openmp  import parse as omp_parse
 from pyccel.parser.syntax.openacc import parse as acc_parse
 
@@ -158,7 +158,7 @@ class SyntaxParser(BasicParser):
 
         Returns
         -------
-        pyccel.ast.basic.Basic
+        pyccel.ast.basic.PyccelAstNode
             The Pyccel AST object.
         """
 
@@ -181,16 +181,55 @@ class SyntaxParser(BasicParser):
         return (self._visit(i) for i in stmt)
 
     def _treat_comment_line(self, line, stmt):
+        """
+        Parse a comment line.
+
+        Parse a comment which fits in a single line. If the comment
+        begins with `#$` then it should contain a header recognised
+        by Pyccel and should be parsed using textx.
+
+        Parameters
+        ----------
+        line : str
+            The comment line.
+        stmt : ast.Ast
+            The comment object in the code. This is useful for raising
+            neat errors.
+
+        Returns
+        -------
+        pyccel.ast.basic.PyccelAstNode
+            The treated object as a Pyccel ast node.
+        """
         if line.startswith('#$'):
             env = line[2:].lstrip()
             if env.startswith('omp'):
-                expr = omp_parse(stmts=line)
+                try:
+                    expr = omp_parse(stmts=line)
+                except TextXSyntaxError as e:
+                    errors.report(f"Invalid OpenMP header. {e.message}",
+                            symbol = stmt, column = e.col,
+                              severity='fatal')
             elif env.startswith('acc'):
-                expr = acc_parse(stmts=line)
+                try:
+                    expr = acc_parse(stmts=line)
+                except TextXSyntaxError as e:
+                    errors.report(f"Invalid OpenACC header. {e.message}",
+                            symbol = stmt, column = e.col,
+                              severity='fatal')
             elif env.startswith('header'):
-                expr = hdr_parse(stmts=line)
-                if isinstance(expr, MetaVariable):
-
+                try:
+                    expr = hdr_parse(stmts=line)
+                except TextXSyntaxError as e:
+                    errors.report(f"Invalid header. {e.message}",
+                            symbol = stmt, column = e.col,
+                              severity='fatal')
+                if isinstance(expr, (MethodHeader, FunctionHeader)):
+                    self.scope.insert_header(expr)
+                    expr = EmptyNode()
+                elif isinstance(expr, AnnotatedPyccelSymbol):
+                    self.scope.insert_symbol(expr.name)
+                elif isinstance(expr, MetaVariable):
                     # a metavar will not appear in the semantic stage.
                     # but can be used to modify the ast
 
@@ -207,6 +246,56 @@ class SyntaxParser(BasicParser):
 
         expr.ast = stmt
         return expr
+
+    def _treat_type_annotation(self, stmt, annotation):
+        """
+        Treat an object passed as a type annotation.
+
+        Ensure that an object that was passed as a type annotation can be
+        recognised in the semantic stage by packing it into a SyntacticTypeAnnotation
+        in the correct way. Also check the syntax of any string type
+        annotations.
+
+        Parameters
+        ----------
+        stmt : ast.Ast
+            The ast node about which any errors should be raised.
+
+        annotation : pyccel.ast.basic.PyccelAstNode
+            A visited object which is describing a type annotation.
+
+        Returns
+        -------
+        SyntacticTypeAnnotation | UnionTypeAnnotation
+            The type annotation.
+        """
+        if isinstance(annotation, (tuple, list)):
+            return UnionTypeAnnotation(*[self._treat_type_annotation(stmt, a) for a in annotation])
+        if isinstance(annotation, (PyccelSymbol, DottedName)):
+            return SyntacticTypeAnnotation(dtypes=[annotation], ranks=[0], orders=[None], is_const=False)
+        elif isinstance(annotation, IndexedElement):
+            return SyntacticTypeAnnotation(dtypes=[annotation], ranks=[len(annotation.indices)], orders=[None], is_const=False)
+        elif isinstance(annotation, LiteralString):
+            try:
+                annotation = types_meta.model_from_str(annotation.python_value)
+            except TextXSyntaxError as e:
+                errors.report(f"Invalid header. {e.message}",
+                        symbol = stmt, column = e.col,
+                        severity='fatal')
+            annot = SyntacticTypeAnnotation.build_from_textx(annotation)
+            if isinstance(stmt, PyccelAstNode):
+                annot.set_fst(stmt.fst)
+            else:
+                annot.set_fst(stmt)
+            return annot
+        elif annotation is Nil():
+            return None
+        elif isinstance(annotation, PyccelBitOr):
+            return UnionTypeAnnotation(*[self._treat_type_annotation(stmt, a) for a in annotation.args])
+        else:
+            errors.report('Invalid type annotation',
+                        symbol = stmt, severity='error')
+            return EmptyNode()
 
     #====================================================
     #                 _visit functions
@@ -242,7 +331,7 @@ class SyntaxParser(BasicParser):
         if hasattr(self, syntax_method):
             self._context.append(stmt)
             result = getattr(self, syntax_method)(stmt)
-            if isinstance(result, Basic) and result.ast is None and isinstance(stmt, ast.AST):
+            if isinstance(result, PyccelAstNode) and result.ast is None and isinstance(stmt, ast.AST):
                 result.ast = stmt
             self._context.pop()
             return result
@@ -368,26 +457,35 @@ class SyntaxParser(BasicParser):
         arguments       = []
         if stmt.args:
             n_expl = len(stmt.args)-len(stmt.defaults)
-            positional_args        = [FunctionDefArgument(PyccelSymbol(a.arg),
-                                            annotation=self._visit(a.annotation))
-                                        for a in stmt.args[:n_expl]]
-            valued_arguments       = [FunctionDefArgument(PyccelSymbol(a.arg),
-                                            annotation=self._visit(a.annotation),
+
+            arguments = []
+            for a in stmt.args[:n_expl]:
+                annotation=self._treat_type_annotation(a, self._visit(a.annotation))
+                new_arg = FunctionDefArgument(AnnotatedPyccelSymbol(a.arg, annotation),
+                                            annotation=annotation)
+                new_arg.set_fst(a)
+                arguments.append(new_arg)
+
+            for a,d in zip(stmt.args[n_expl:], stmt.defaults):
+                annotation=self._treat_type_annotation(a, self._visit(a.annotation))
+                new_arg = FunctionDefArgument(AnnotatedPyccelSymbol(a.arg, annotation),
+                                            annotation=annotation,
                                             value = self._visit(d))
-                                        for a,d in zip(stmt.args[n_expl:],stmt.defaults)]
-            arguments              = positional_args + valued_arguments
-            self.scope.insert_symbols(PyccelSymbol(a.arg) for a in stmt.args)
+                new_arg.set_fst(a)
+                arguments.append(new_arg)
 
         if stmt.kwonlyargs:
             for a,d in zip(stmt.kwonlyargs,stmt.kw_defaults):
-                annotation = self._visit(a.annotation)
+                annotation=self._treat_type_annotation(a, self._visit(a.annotation))
                 val = self._visit(d) if d is not None else d
-                arg = FunctionDefArgument(PyccelSymbol(a.arg),
+                arg = FunctionDefArgument(AnnotatedPyccelSymbol(a.arg, annotation),
                             annotation=annotation,
                             value=val, kwonly=True)
+                arg.set_fst(a)
 
                 arguments.append(arg)
-                self.scope.insert_symbol(a.arg)
+
+        self.scope.insert_symbols(a.var for a in arguments)
 
         return arguments
 
@@ -621,14 +719,13 @@ class SyntaxParser(BasicParser):
 
         name = PyccelSymbol(self._visit(stmt.name))
         self.scope.insert_symbol(name)
-        name = name.replace("'", '')
+
+        headers = self.scope.find(name, 'headers')
 
         scope = self.create_new_function_scope(name)
 
         arguments    = self._visit(stmt.args)
 
-        global_vars  = []
-        headers      = []
         template    = {}
         is_pure      = False
         is_elemental = False
@@ -637,31 +734,7 @@ class SyntaxParser(BasicParser):
         imports      = []
         doc_string   = None
 
-        def fill_types(ls):
-            container = []
-            for arg in ls:
-                if isinstance(arg, FunctionCallArgument):
-                    arg = arg.value
-
-                if isinstance(arg, PyccelSymbol):
-                    container.append(arg)
-                elif isinstance(arg, LiteralString):
-                    arg = str(arg)
-                    arg = arg.strip("'").strip('"')
-                    container.append(arg)
-                else:
-                    msg = f'Invalid argument of type {type(arg)} passed to types decorator'
-                    errors.report(msg,
-                                bounding_box = (stmt.lineno, stmt.col_offset),
-                                severity='error')
-            return container
-
         decorators = {}
-
-        # add the decorator @types if the arguments are annotated
-        annotated_args = []
-        for a in arguments:
-            annotated_args.append(a.annotation)
 
         for d in self._visit(stmt.decorator_list):
             tmp_var = d if isinstance(d, PyccelSymbol) else d.funcdef
@@ -673,126 +746,12 @@ class SyntaxParser(BasicParser):
         if 'types' in decorators:
             warnings.warn("The @types decorator will be removed in a future version of Pyccel. Please use type hints. The @template decorator can be used to specify multiple types", FutureWarning)
 
-        if all(not isinstance(a, Nil) for a in annotated_args):
-            if stmt.returns:
-                returns = FunctionCallArgument(self._visit(stmt.returns), keyword='results')
-                annotated_args.append(returns)
-            decorators.setdefault('types', []).append(FunctionCall('types', annotated_args))
-
-        if 'bypass' in decorators:
-            return EmptyNode()
-
         if 'stack_array' in decorators:
             decorators['stack_array'] = tuple(str(b.value) for a in decorators['stack_array']
                 for b in a.args)
 
         if 'allow_negative_index' in decorators:
             decorators['allow_negative_index'] = tuple(str(b.value) for a in decorators['allow_negative_index'] for b in a.args)
-        template['template_dict'] = {}
-        # extract the templates
-        if 'template' in decorators:
-            for comb_types in decorators['template']:
-                cache.clear_cache()
-                types = []
-                if len(comb_types.args) != 2:
-                    msg = 'Number of Arguments provided to the template decorator is not valid'
-                    errors.report(msg,
-                                    symbol = comb_types,
-                                    bounding_box = (stmt.lineno, stmt.col_offset),
-                                    severity='error')
-
-                for i in comb_types.args:
-                    if i.has_keyword and i.keyword not in ('name', 'types'):
-                        msg = 'Argument provided to the template decorator is not valid'
-                        errors.report(msg,
-                                        symbol = comb_types,
-                                        bounding_box = (stmt.lineno, stmt.col_offset),
-                                        severity='error')
-                if all(i.has_keyword for i in comb_types.args):
-                    tp_name, ls = (comb_types.args[0].value, comb_types.args[1].value) if\
-                            comb_types.args[0].keyword == 'name' else\
-                            (comb_types.args[1].value, comb_types.args[0].value)
-                else:
-                    tp_name = comb_types.args[0].value
-                    ls = comb_types.args[1].value
-                try:
-                    tp_name = str(tp_name)
-                    ls = ls if isinstance(ls, PythonTuple) else list(ls)
-                except TypeError:
-                    msg = 'Argument provided to the template decorator is not valid'
-                    errors.report(msg,
-                                    symbol = comb_types,
-                                    bounding_box = (stmt.lineno, stmt.col_offset),
-                                    severity='fatal')
-
-                types = fill_types(ls)
-
-                txt  = '#$ header template ' + tp_name
-                txt += '(' + '|'.join(types) + ')'
-                if tp_name in template['template_dict']:
-                    msg = f'The template "{tp_name}" is duplicated'
-                    errors.report(msg,
-                                bounding_box = (stmt.lineno, stmt.col_offset),
-                                severity='warning')
-                # Make templates decorator dict accessible from decorators dict
-                template['template_dict'][tp_name] = hdr_parse(stmts=txt)
-            # Make template decorator list accessible from decorators dict
-            template['decorator_list'] = decorators['template']
-            decorators['template'] = template
-
-        if not template['template_dict']:
-            decorators['template'] = None
-        # extract the types to construct a header
-        if 'types' in decorators:
-            for comb_types in decorators['types']:
-
-                cache.clear_cache()
-                results = []
-                ls = comb_types.args
-
-                if len(ls) > 0 and ls[-1].has_keyword:
-                    arg_name = ls[-1].keyword
-                    if not arg_name == 'results':
-                        msg = f'Argument "{arg_name}" provided to the types decorator is not valid'
-                        errors.report(msg,
-                                    symbol = comb_types,
-                                    bounding_box = (stmt.lineno, stmt.col_offset),
-                                    severity='error')
-                    else:
-                        container = ls[-1].value
-                        container = container if isinstance(container, PythonTuple) else [container]
-                        results = fill_types(container)
-                    types = fill_types(ls[:-1])
-                else:
-                    types = fill_types(ls)
-
-                txt  = '#$ header '
-                if len(self._context) > 1 and isinstance(self._context[-2], ast.ClassDef):
-                    txt += 'method '
-                txt += name + '(' + ','.join(types) + ')'
-
-                if results:
-                    txt += ' results(' + ','.join(results) + ')'
-
-                header = hdr_parse(stmts=txt)
-                headers += [header]
-
-        body = stmt.body
-
-        if 'sympy' in decorators:
-            # TODO maybe we should run pylint here
-            stmt.decorators.pop()
-            func = SympyFunction(name, arguments, [], [str(stmt)])
-            func.ast = stmt
-            self.insert_function(func)
-            return EmptyNode()
-
-        else:
-            body = self._visit(body)
-        if len(body) > 0 and isinstance(body[0], CommentBlock):
-            doc_string = body[0]
-            doc_string.header = ''
-            body = body[1:]
 
         if 'pure' in decorators:
             is_pure = True
@@ -811,6 +770,171 @@ class SyntaxParser(BasicParser):
         if 'inline' in decorators:
             is_inline = True
 
+        template['template_dict'] = {}
+        # extract the templates
+        if 'template' in decorators:
+            for template_decorator in decorators['template']:
+                dec_args = template_decorator.args
+                if len(dec_args) != 2:
+                    msg = 'Number of Arguments provided to the template decorator is not valid'
+                    errors.report(msg, symbol = template_decorator,
+                                    severity='error')
+
+                if any(i.keyword not in (None, 'name', 'types') for i in dec_args):
+                    errors.report('Argument provided to the template decorator is not valid',
+                                    symbol = template_decorator, severity='error')
+
+                if dec_args[0].has_keyword and dec_args[0].keyword != 'name':
+                    type_name = dec_args[1].value.python_value
+                    type_descriptors = dec_args[0].value
+                else:
+                    type_name = dec_args[0].value.python_value
+                    type_descriptors = dec_args[1].value
+
+                if not isinstance(type_descriptors, (PythonTuple, PythonList)):
+                    type_descriptors = PythonTuple(type_descriptors)
+
+                if type_name in template['template_dict']:
+                    errors.report(f'The template "{type_name}" is duplicated',
+                                symbol = template_decorator, severity='warning')
+
+                possible_types = self._treat_type_annotation(template_decorator, type_descriptors.args)
+
+                # Make templates decorator dict accessible from decorators dict
+                template['template_dict'][type_name] = possible_types
+
+            # Make template decorator list accessible from decorators dict
+            template['decorator_list'] = decorators['template']
+            decorators['template'] = template
+
+        if not template['template_dict']:
+            decorators['template'] = None
+
+        argument_annotations = [a.annotation for a in arguments]
+        result_annotation = self._treat_type_annotation(stmt, self._visit(stmt.returns))
+
+        #---------------------------------------------------------------------------------------------------------
+        #                   To remove when headers are deprecated
+        #---------------------------------------------------------------------------------------------------------
+        if headers:
+            warnings.warn("Support for specifying types via headers will be removed in a " +
+                          "future version of Pyccel. Please use type hints. The @template " +
+                          "decorator can be used to specify multiple types. See the " +
+                          "documentation at " +
+                          "https://github.com/pyccel/pyccel/blob/devel/docs/quickstart.md#type-annotations " +
+                          "for examples.", FutureWarning)
+            if any(a is not None for a in argument_annotations):
+                errors.report("Type annotations and type specification via headers should not be mixed",
+                        symbol=stmt, severity='error')
+
+            for i, _ in enumerate(argument_annotations):
+                argument_annotations[i] = UnionTypeAnnotation()
+            if result_annotation is not None:
+                errors.report("Type annotations and type specification via headers should not be mixed",
+                            symbol=stmt, severity='error')
+
+            n_results = 0
+
+            for head in headers:
+                if len(head.dtypes) != len(argument_annotations):
+                    errors.report(f"Wrong number of types in header for function {name}",
+                            severity='error', symbol=stmt)
+                else:
+                    for i,arg in enumerate(head.dtypes):
+                        argument_annotations[i].add_type(arg)
+                if head.results:
+                    if result_annotation is None:
+                        result_annotation = head.results
+                    else:
+                        if len(result_annotation) != len(head.results):
+                            errors.report("Different length results in headers.",
+                                    severity='error', symbol=stmt)
+                        else:
+                            result_annotation = tuple(UnionTypeAnnotation(r, *getattr(t, 'type_list', [t])) \
+                                                        for r,t in zip(head.results, result_annotation))
+                    n_results += 1
+
+            if n_results and n_results != len(headers):
+                errors.report("Results have only been provided for some of the headers. The types of the results will all be chosen from the provided types. This may result in unexpected result types.",
+                        severity='warning', symbol=stmt)
+
+        # extract the types to construct a header
+        if 'types' in decorators:
+            if any(a is not None for a in argument_annotations):
+                errors.report("Type annotations and type specification via headers should not be mixed",
+                        symbol=stmt, severity='error')
+
+            for i, _ in enumerate(argument_annotations):
+                argument_annotations[i] = UnionTypeAnnotation()
+            n_results = 0
+
+            for types_decorator in decorators['types']:
+                type_args = types_decorator.args
+
+                args = [a for a in type_args if not a.has_keyword]
+                kwargs = [a for a in type_args if a.has_keyword]
+
+                if len(kwargs) > 1:
+                    errors.report('Too many keyword arguments passed to @types decorator',
+                                symbol = types_decorator,
+                                bounding_box = (stmt.lineno, stmt.col_offset),
+                                severity='error')
+                elif kwargs:
+                    if kwargs[0].keyword != 'results':
+                        errors.report('Wrong keyword argument passed to @types decorator',
+                                    symbol = types_decorator,
+                                    bounding_box = (stmt.lineno, stmt.col_offset),
+                                    severity='error')
+                    annots = self._treat_type_annotation(kwargs[0], kwargs[0].value.args)
+                    if result_annotation is None:
+                        result_annotation = annots
+                    else:
+                        if len(result_annotation) != len(annots):
+                            errors.report("Different length results in headers.",
+                                    severity='error', symbol=stmt)
+                        else:
+                            result_annotation = tuple(UnionTypeAnnotation(r, *getattr(t, 'type_list', [t])) \
+                                                        for r,t in zip(annots, result_annotation))
+                    n_results += 1
+
+                if len(args) != len(argument_annotations):
+                    errors.report(f"Wrong number of types in header for function {name}",
+                            severity='error', symbol=stmt)
+                else:
+                    for i,arg in enumerate(args):
+                        argument_annotations[i].add_type(self._treat_type_annotation(arg, arg.value))
+
+            if n_results and n_results != len(decorators['types']):
+                errors.report("Results have only been provided for some of the types decorators. The types of the results will all be chosen from the provided types. This may result in unexpected result types.",
+                        severity='warning', symbol=stmt)
+
+        #---------------------------------------------------------------------------------------------------------
+        #                   End of : To remove when headers are deprecated
+        #---------------------------------------------------------------------------------------------------------
+
+        # Repack AnnotatedPyccelSymbols to insert argument_annotations from headers or types decorators
+        arguments = [FunctionDefArgument(AnnotatedPyccelSymbol(a.var.name, annot), annotation=annot, value=a.value, kwonly=a.is_kwonly)
+                           for a, annot in zip(arguments, argument_annotations)]
+
+        body = stmt.body
+
+        if 'sympy' in decorators:
+            # TODO maybe we should run pylint here
+            stmt.decorators.pop()
+            func = SympyFunction(name, arguments, [], [str(stmt)])
+            func.ast = stmt
+            self.insert_function(func)
+            return EmptyNode()
+
+        else:
+            body = self._visit(body)
+
+        # Collect docstring
+        if len(body) > 0 and isinstance(body[0], CommentBlock):
+            doc_string = body[0]
+            doc_string.header = ''
+            body = body[1:]
+
         body = CodeBlock(body)
 
         returns = [i.expr for i in body.get_attribute_nodes(Return,
@@ -825,7 +949,10 @@ class SyntaxParser(BasicParser):
 
         local_symbols = self.scope.local_used_symbols
 
-        for r in zip(*returns):
+        if result_annotation and not isinstance(result_annotation, tuple):
+            result_annotation = [result_annotation]
+
+        for i,r in enumerate(zip(*returns)):
             r0 = r[0]
 
             pyccel_symbol  = isinstance(r0, PyccelSymbol)
@@ -837,7 +964,11 @@ class SyntaxParser(BasicParser):
             else:
                 result_name, result_counter = self.scope.get_new_incremented_symbol('Out', result_counter)
 
-            results.append(FunctionDefResult(result_name))
+            if result_annotation:
+                result_name = AnnotatedPyccelSymbol(result_name, annotation = result_annotation[i])
+
+            results.append(FunctionDefResult(result_name, annotation = result_annotation))
+            results[-1].set_fst(stmt)
 
         self.exit_function_scope()
 
@@ -847,13 +978,11 @@ class SyntaxParser(BasicParser):
                arguments,
                results,
                body,
-               global_vars=global_vars,
                is_pure=is_pure,
                is_elemental=is_elemental,
                is_private=is_private,
                imports=imports,
                decorators=decorators,
-               headers=headers,
                doc_string=doc_string,
                scope=scope)
 
@@ -911,19 +1040,24 @@ class SyntaxParser(BasicParser):
 
     def _visit_Attribute(self, stmt):
         val  = self._visit(stmt.value)
-        if self._in_lhs_assign:
-            self.scope.insert_symbol(stmt.attr)
         attr = PyccelSymbol(stmt.attr)
-        return DottedName(val, attr)
+        dotted = DottedName(val, attr)
+        if self._in_lhs_assign:
+            self.scope.insert_symbol(dotted)
+        return dotted
 
 
     def _visit_Call(self, stmt):
 
         args = []
         if stmt.args:
-            args += [FunctionCallArgument(self._visit(a)) for a in stmt.args]
+            args += [FunctionCallArgument(self._visit(a), fst=a) for a in stmt.args]
         if stmt.keywords:
-            args += self._visit(stmt.keywords)
+            kwargs = self._visit(stmt.keywords)
+            for k, a in zip(kwargs, stmt.keywords):
+                k.set_fst(a)
+
+            args += kwargs
 
         if len(args) == 0:
             args = ()
