@@ -1,9 +1,9 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 from pyccel.decorators import template
 
-@template('z', types=['int', 'real'])
-@template('y', types=['int', 'real'])
-def multi_tmplt_1(x : 'z', y : 'z', z : 'y'):
+@template('a', types=['int', 'real'])
+@template('b', types=['int', 'real'])
+def multi_tmplt_1(x : 'a', y : 'a', z : 'b'):
     """Tests Interfaces"""
     return x + y + z
 
