@@ -12,20 +12,21 @@ class Point(object):
         self.x = self.x + a
         self.y = self.y + b
 
-p = Point(0.0, 0.0)
-x=p.x
-p.x=x
-a = p.x
-a = p.x - 2
-a = 2 * p.x - 2
-a = 2 * (p.x + 6) - 2
+if __name__ == '__main__':
+    p = Point(0.0, 0.0)
+    x=p.x
+    p.x=x
+    a = p.x
+    a = p.x - 2
+    a = 2 * p.x - 2
+    a = 2 * (p.x + 6) - 2
 
-p.y = a + 5
-p.y = p.x + 5
+    p.y = a + 5
+    p.y = p.x + 5
 
-p.translate(1.0, 2.0)
+    p.translate(1.0, 2.0)
 
-print(p.x, p.y)
-print(a)
+    print(p.x, p.y)
+    print(a)
 
-del p
+    del p
